Methods ADD Prof (I)

+++ src/main/java/com/mundomuebles/mundo_muebles_fd/infrastructure/adapters/PostgresCategoryRepository.java
@@ -2,6 +2,7 @@ package com.mundomuebles.mundo_muebles_fd.infrastructure.adapters;
 
 import com.mundomuebles.mundo_muebles_fd.application.CategoryAble;
 import com.mundomuebles.mundo_muebles_fd.domain.CategoryDTO;
+import com.mundomuebles.mundo_muebles_fd.exception.AppException;
 import com.mundomuebles.mundo_muebles_fd.infrastructure.repositories.Category;
 import com.mundomuebles.mundo_muebles_fd.infrastructure.repositories.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,14 +26,43 @@ public class PostgresCategoryRepository implements CategoryAble {
 
     @Override
 
-    public CategoryDTO update(CategoryDTO categoryDTO) {
-        return categoryRepository.save(new Category(categoryDTO)).toCategoryDTO();
+    public CategoryDTO update(CategoryDTO categoryDTO)throws AppException{
+        if (categoryRepository.existsById(categoryDTO.getId())){
+            try{
+                return
+                categoryRepository.save(new Category(categoryDTO)).toCategoryDTO();
+
+            }
+            catch(Exception e) {
+                throw new AppException("No se puede actualizar el registro");
+            }
+
+
+        }
+        else {
+            throw new AppException("El registro no existe");
+
+        }
     }
 
     @Override
-    public boolean delete(int id) {
-        categoryRepository.deleteById(id);
-        return true;
+    public boolean delete(int id) throws AppException {
+        if (categoryRepository.existsById(id)){
+            try{
+                categoryRepository.deleteById(id);
+                return true;
+            }
+            catch(Exception e) {
+                throw new AppException("El registro no se puede eliminar");
+            }
+
+
+        }
+        else {
+            throw new AppException("El registro a eliminar no existe");
+
+        }
+
     }
 
     @Override
