+++ src/main/java/com/mundomuebles/mundo_muebles_fd/infrastructure/adapters/PostgresCreditRepository.java
@@ -2,6 +2,7 @@ package com.mundomuebles.mundo_muebles_fd.infrastructure.adapters;
 
 import com.mundomuebles.mundo_muebles_fd.application.CreditAble;
 import com.mundomuebles.mundo_muebles_fd.domain.CreditDTO;
+import com.mundomuebles.mundo_muebles_fd.exception.AppException;
 import com.mundomuebles.mundo_muebles_fd.infrastructure.repositories.Credit;
 import com.mundomuebles.mundo_muebles_fd.infrastructure.repositories.CreditRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,14 +25,40 @@ public class PostgresCreditRepository implements CreditAble {
     }
 
     @Override
-    public CreditDTO update(CreditDTO creditDTO) {
-        return creditRepository.save(new Credit(creditDTO)).toCreditDTO();
+    public CreditDTO update(CreditDTO creditDTO)throws AppException{
+        if (creditRepository.existsById(creditDTO.getId())){
+            try{
+                return
+                        creditRepository.save(new Credit(creditDTO)).toCreditDTO();
+
+            }
+            catch(Exception e) {
+                throw new AppException("No se puede actualizar el registro");
+            }
+
+
+        }
+        else {
+            throw new AppException("El registro no existe");
+
+        }
     }
 
     @Override
-    public boolean delete(int id) {
-        creditRepository.deleteById(id);
-        return true;
+    public boolean delete(int id) throws AppException {
+        if (creditRepository.existsById(id)){
+            try{
+                creditRepository.deleteById(id);
+                return true;
+            }
+            catch(Exception e) {
+                throw new AppException("El registro no se puede eliminar");
+            }
+        }
+        else {
+            throw new AppException("El registro a eliminar no existe");
+
+        }
     }
 
     @Override
