+++ src/main/java/com/mundomuebles/mundo_muebles_fd/infrastructure/adapters/PostgresPurchaserRepository.java
@@ -2,6 +2,7 @@ package com.mundomuebles.mundo_muebles_fd.infrastructure.adapters;
 
 import com.mundomuebles.mundo_muebles_fd.application.PurchaserAble;
 import com.mundomuebles.mundo_muebles_fd.domain.PurchaserDTO;
+import com.mundomuebles.mundo_muebles_fd.exception.AppException;
 import com.mundomuebles.mundo_muebles_fd.infrastructure.repositories.Purchaser;
 import com.mundomuebles.mundo_muebles_fd.infrastructure.repositories.PurchaserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,14 +26,41 @@ public class PostgresPurchaserRepository implements PurchaserAble {
     }
 
     @Override
-    public PurchaserDTO update(PurchaserDTO purchaserDTO) {
-        return purchaserRepository.save(new Purchaser(purchaserDTO)).toPurchaserDTO();
+    public PurchaserDTO update(PurchaserDTO purchaserDTO)throws AppException{
+        if (purchaserRepository.existsById(purchaserDTO.getId())){
+            try{
+                return
+                        purchaserRepository.save(new Purchaser(purchaserDTO)).toPurchaserDTO();
+
+            }
+            catch(Exception e) {
+                throw new AppException("No se puede actualizar el registro");
+            }
+
+
+        }
+        else {
+            throw new AppException("El registro no existe");
+
+        }
     }
 
     @Override
-    public boolean delete(BigInteger id) {
-        purchaserRepository.deleteById(id);
-        return true;
+    public boolean delete(BigInteger id) throws AppException {
+        if (purchaserRepository.existsById(id)){
+            try{
+                purchaserRepository.deleteById(id);
+                return true;
+            }
+            catch(Exception e) {
+                throw new AppException("El registro no se puede eliminar");
+            }
+
+        }
+        else {
+            throw new AppException("El registro a eliminar no existe");
+
+        }
     }
 
     @Override
